Added app settings
