+++ .vscode/launch.json
@@ -12,9 +12,9 @@
             "mode": "auto",
             "program": "server.go",
             "env": {
-                "ADMIN_USER": "admin",
-		        "ADMIN_PASS": "pass",
-		        "ADMIN_TOKEN": "token"
+                "TTTSERVER_ADMIN_USER": "admin",
+		        "TTTSERVER_ADMIN_PASS": "pass",
+		        "TTTSERVER_ADMIN_TOKEN": "token"
             }
         }
     ]
