+++ internal/server/server.go
@@ -4,42 +4,17 @@ import (
 	"encoding/json"
 	"errors"
 	"net/http"
-	"os"
 	"sync"
+
+	"github.com/the-developer-guy/tic-tac-multiplayer/internal"
 )
-
-type AppConfig struct {
-	AdminUser     string
-	AdminPassword string
-	AdminToken    string
-}
-
-func loadConfig() {
-
-	ac := AppConfig{
-		AdminUser:     os.Getenv("ADMIN_USER"),
-		AdminPassword: os.Getenv("ADMIN_PASS"),
-		AdminToken:    os.Getenv("ADMIN_TOKEN"),
-	}
-
-	if ac.AdminUser == "" {
-		panic("Missing admin username from config")
-	}
-	if ac.AdminPassword == "" {
-		panic("Missing admin password from config")
-	}
-	if ac.AdminToken == "" {
-		panic("Missing admin token from config")
-	}
-}
 
 type TicTacToeServer struct {
 	Lobbies     map[string]*Lobby
 	lobbiesLock sync.Mutex
 }
 
-func NewTicTacToeServer() *TicTacToeServer {
-	loadConfig()
+func NewTicTacToeServer(ac *internal.AppConfig) *TicTacToeServer {
 	ttts := TicTacToeServer{
 		Lobbies: make(map[string]*Lobby),
 	}
